feat(workflow): migrate manual, scheduled, date-time and format to the catalog
Third catalog batch (8 of 38 types migrated). Data halves move to
catalog/nodes/{manual,scheduled,date-time,format}.ts; each web schema.ts
shrinks to its defineFromManifest merge site plus back-compat re-exports;
zero diff under any node.tsx / panel.tsx / trace-renderer.tsx.

Along the way:
- The defaults-parse discipline fired again: date-time's schema required
  inputDate min(1) while its defaults ship without one (the user picks it in
  the panel). Same resolution as var-assign's empty starter row: the schema
  allows the empty persisted state, completeness stays in the validator.
- The coverage fixture now also supplies `title` (node-factory falls back to
  definition.displayName), which scheduled's title-less defaults surfaced.
- Engine sweep deletes date-time-processor's four verbatim enum
  re-declarations (DateTimeOperation, TimeUnit, DateFormatType,
  ParseDateFormatType) - the catalog is now their single source.
- The parity suite's builderDeclaredKeys reader learns that a migrated
  node's data interface lives in the catalog file, not the web types.ts
  narrowing wrapper - without this, format's 18 per-operation config keys
  read as engine-only.
- format's extractVariables (incl. the bare-dotted-path extractVariableRefs
  scan) and scheduled's move into their manifests; format's per-operation
  config types were already lib-side in workflow-engine/constants.
